# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/google/callback": {
      "get": {
        "operationId": "AppController_googleAuthRedirect",
        "summary": "Google Authentication Callback",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/store/order/{id}": {
      "get": {
        "operationId": "StoreController_findOne",
        "summary": "Get an order by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "tags": [
          "store"
        ]
      },
      "delete": {
        "operationId": "StoreController_remove",
        "summary": "Delete an order by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted"
          }
        },
        "tags": [
          "store"
        ]
      }
    },
    "/store/order": {
      "post": {
        "operationId": "StoreController_create",
        "summary": "Create a new order",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "tags": [
          "store"
        ]
      }
    }
  },
  "info": {
    "title": "Sport Shop API",
    "description": "The Sport Shop API documentation",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the user"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "The last name of the user"
          },
          "picture": {
            "type": "string",
            "example": "http://example.com/avatar.jpg",
            "description": "The profile picture of the user"
          },
          "accessToken": {
            "type": "string",
            "example": "token123",
            "description": "The access token of the user"
          },
          "role": {
            "type": "string",
            "example": "user",
            "description": "The role of the user",
            "default": "user"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "picture",
          "accessToken",
          "role"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the order"
          },
          "sporting_goodsId": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the sporting goods item"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "The quantity of the sporting goods item"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "description": "The status of the order"
          },
          "orderDate": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01T00:00:00Z",
            "description": "The date the order was placed"
          }
        },
        "required": [
          "id",
          "sporting_goodsId",
          "quantity",
          "status",
          "orderDate"
        ]
      }
    }
  }
}